refactor: extract models from dto to model package
