+++ src/main/kotlin/ecommerce/repository/JdbcCartRepository.kt
@@ -1,6 +1,6 @@
 package ecommerce.repository
 
-import ecommerce.dto.CartItem
+import ecommerce.model.CartItem
 import org.springframework.jdbc.core.JdbcTemplate
 import org.springframework.jdbc.core.RowMapper
 import org.springframework.stereotype.Repository
