+++ src/test/kotlin/ecommerce/dto/MemberTest.kt
@@ -1,5 +1,6 @@
 package ecommerce.dto
 
+import ecommerce.model.Member
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
