+++ src/test/kotlin/ecommerce/dto/ProductTest.kt
@@ -1,5 +1,6 @@
 package ecommerce.dto
 
+import ecommerce.model.Product
 import ecommerce.utiles.Constants
 import jakarta.validation.Validation
 import jakarta.validation.Validator
