+++ src/test/kotlin/ecommerce/service/CartServiceTest.kt
@@ -1,6 +1,6 @@
 package ecommerce.service
 
-import ecommerce.dto.CartItem
+import ecommerce.model.CartItem
 import ecommerce.repository.CartRepository
 import io.mockk.every
 import io.mockk.mockk
